light: current-frame flush origin + double light positions
Companion to irl-core cf9cff6: second renderWorld inject after Camera.update
calls FramePipeline.uploadIfPending() (SSBO origin = this frame's eye), and
LightDriver passes PlacedLight coordinates as doubles into the registry.

+++ src/client/java/org/qualet/irlredactor/light/LightDriver.java
@@ -117,7 +117,7 @@ public final class LightDriver
     private static void emitPoint(PlacedLight l)
     {
         LightRegistry.registerPoint(
-            (float) l.x, (float) l.y, (float) l.z,
+            l.x, l.y, l.z,
             l.r, l.g, l.b,
             l.intensity, l.radius,
             l.entitiesOnly, l.blocksOnly,
@@ -142,7 +142,7 @@ public final class LightDriver
         float cookieFlags = l.cookieInvert ? 1F : 0F;
 
         LightRegistry.registerSpot(
-            (float) l.x, (float) l.y, (float) l.z,
+            l.x, l.y, l.z,
             dx, dy, dz,
             l.r, l.g, l.b,
             l.intensity, l.range, cosOuter, cosInner,
